pyproject.toml
[project]
name = "optionstrat-backend"
version = "0.1.0"
description = "Backend API for Options Scenario Analysis & Exploration App"
readme = "src/backend/README.md"
requires-python = ">=3.13.11,<4.0"
authors = [{ name = "Jeff Coggshall", email = "thenextlocalminima@gmail..com" }]
dependencies = [
  "fastapi>=0.109.0",
  "starlette>=0.46.0",
  "uvicorn>=0.27.0",
  "sqlalchemy>=2.0.25",
  "pydantic>=2.5.3",
  "quantlib>=1.32",
  "yfinance>=0.2.35",
  "numpy>=1.26.3",
  "pandas>=2.1.4",
  "matplotlib>=3.8.0",
  "requests>=2.31.0",
  "redis>=5.0.1",
  "python-dotenv>=1.0.0",
  "markdown-it-py>=3.0.0",
  "plotly>=6.3.0",
  "pytest>=8.4.2",
  "httpx>=0.28.1",
  "scikit-learn>=1.8.0",
]

[project.scripts]
security-scan = "app.security_scan.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.4",
  "pytest-cov>=4.1.0",
  "black>=23.7.0",
  "isort>=5.12.0",
  "mypy>=1.5.1",
  "flake8>=6.1.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["quantlib.*"]
ignore_missing_imports = true 

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = "src/backend"
module-name = "app"
